# Add options to support `Profile Guided Optimization`