--- a/libs/libcommon/include/common/config_common.h.in
+++ b/libs/libcommon/include/common/config_common.h.in
@@ -14,3 +14,7 @@
 #cmakedefine01 USE_UNWIND
 #cmakedefine01 USE_LLVM_LIBUNWIND
 #cmakedefine01 ENABLE_THINLTO
+#cmakedefine01 ENABLE_LLVM_PGO
+#cmakedefine01 ENABLE_LLVM_PROFILE_INSTR
+#cmakedefine01 ENABLE_LLVM_PGO_USE_SAMPLE
+#cmakedefine01 USE_LLVM_FDO