### What problem does this PR solve?

Issue Number: ref #4909

### What is changed and how it works?

doc: https://solotzg.github.io/2022/06/01/system/cpp-compile-optimize/#PGO%EF%BC%88Profile-Guided-Optimization%EF%BC%89

[Profile Guided Optimization](https://en.wikipedia.org/wiki/Profile-guided_optimization) aka `Feedback Directed Optimization` make it possible to achieve better optimization by profile information. This PR supports profile guided optimization with two different kinds of profiling:
- [Using Sampling Profilers](https://clang.llvm.org/docs/UsersManual.html#using-sampling-profilers): 
  1. add cmake option `ENABLE_LLVM_PGO` and `ENABLE_LLVM_PGO_USE_SAMPLE`, then recompile.
  3. run the executable under a sampling profiler by [Linux Perf profiler](https://perf.wiki.kernel.org/)
  4. convert the collected profile data to LLVM’s sample profile format by [AutoFDO](https://github.com/google/autofdo)
  5. add env var `TIFLASH_LLVM_PROFDATA=xxxx` for recompiling.
- [Profiling with Instrumentation](https://clang.llvm.org/docs/UsersManual.html#profiling-with-instrumentation)
  1. add cmake option `ENABLE_LLVM_PROFILE_INSTR` to generate instrumented code and recompile.
  6. set env `LLVM_PROFILE_FILE=xxxx` then start program and run benchmark to collect profile data
  7. `profdata merge -output=xxxx xxxx.profraw`
  8. add cmake option `ENABLE_LLVM_PGO` and env var `TIFLASH_LLVM_PROFDATA=xxxx` for recompiling

`Profiling with Instrumentation` will harm performance, and CAN NOT collect the real online profile info. So it's recommended to use the sampling way. 

#### Benchmark-10G

env:
- 1 tiflash, 3 pd, 3 tikv, 1 tidb
- data: tpch-10
- limit tiflash cpu usage up to 5 cores(500%) by cgroup.
- original commit: 5b61ae70550624d3bf0b5ca6bac89013ed5a6a4b

workload:

- go-tpc: https://github.com/pingcap/go-tpc/blob/master/tpch/query.go
- use tpch queries to sample profilers and verify

```
git clone git@github.com:pingcap/go-tpc.git
cd go-tpc
make build
bin/go-tpc tpch run --queries "q1,q2,q3,q4,q6,q7,q8,q9,q10,q11,q12,q13,q14,q15,q16,q17,q18,q19,q20,q21,q22" --host {} -P {} --db {}
```

steps:

1. compile TiFlash with cmake option `-DENABLE_LLVM_PGO=ON -DENABLE_LLVM_PGO_USE_SAMPLE=ON`
2. compile https://github.com/google/autofdo and get binary `create_llvm_prof` for converting perf data to llvm profile data
3. start TiFlash process and get `<pid>`
4. prepare workload scripts file
5. run `python3 perf-tpch.py --perf --pid <pid> --workload <workload-scripts-path> --convert-llvm --convert-tool <create_llvm_prof-path> --binary <tiflash-bianry-path>`
6. get llvm perf file(`tiflash.llvm.code.prof` by default)
7. compile TiFlash with env `TIFLASH_LLVM_PROFDATA=<output-perf-file>` and cmake option `-DENABLE_LLVM_PGO=ON -DENABLE_LLVM_PGO_USE_SAMPLE=ON`
8. re-run workload and compare result

Time Cost(s) | original | LTO | FDO+LTO |   | FDO+LTO : LTO | FDO+LTO : original
-- | -- | -- | -- | -- | -- | --
Q1 | 6.32 | 6.07 | 5.57 |   | 8.98% | 13.46%
Q2 | 2.99 | 2.99 | 2.92 |   | 2.40% | 2.40%
Q3 | 2.79 | 2.65 | 2.65 |   | 0.00% | 5.28%
Q4 | 1.91 | 2.05 | 1.85 |   | 10.81% | 3.24%
Q6 | 0.91 | 0.91 | 0.84 |   | 8.33% | 8.33%
Q7 | 2.38 | 2.32 | 2.32 |   | 0.00% | 2.59%
Q8 | 4.8 | 4.73 | 4.73 |   | 0.00% | 1.48%
Q9 | 16.81 | 16.54 | 16.48 |   | 0.36% | 2.00%
Q10 | 3.72 | 3.72 | 3.66 |   | 1.64% | 1.64%
Q11 | 0.5 | 0.5 | 0.5 |   | 0.00% | 0.00%
Q12 | 1.98 | 1.91 | 1.85 |   | 3.24% | 7.03%
Q13 | 4.66 | 4.6 | 4.33 |   | 6.24% | 7.62%
Q14 | 1.04 | 1.11 | 0.97 |   | 14.43% | 7.22%
Q15 | 2.05 | 1.98 | 2.11 |   | -6.16% | -2.84%
Q16 | 1.04 | 1.04 | 0.97 |   | 7.22% | 7.22%
Q17 | 5.67 | 5.8 | 5.67 |   | 2.29% | 0.00%
Q18 | 8.62 | 8.41 | 7.99 |   | 5.26% | 7.88%
Q19 | 3.12 | 3.05 | 3.05 |   | 0.00% | 2.30%
Q20 | 1.58 | 1.58 | 1.64 |   | -3.66% | -3.66%
Q21 | 2.99 | 2.85 | 2.89 |   | -1.38% | 3.46%
Q22 | 0.64 | 0.64 | 0.5 |   | 28.00% | 28.00%

----

Updated in 2022-07-08
- Change original commit since: 97342db2a1dd705b1bd98f3481d51f48a4f0300d

Time Cost(s) | LTO | FDO+LTO | FDO+LTO : LTO
-- | -- | -- | --
Q1 | 5.47 | 5.07 | 7.89%
Q2 | 3.12 | 2.92 | 6.85%
Q3 | 2.72 | 2.58 | 5.43%
Q4 | 2.92 | 2.92 | 0.00%
Q6 | 0.97 | 0.84 | 15.48%
Q7 | 3.59 | 3.52 | 1.99%
Q8 | 5.34 | 5.67 | -5.82%
Q9 | 15.8 | 15.2 | 3.95%
Q10 | 3.72 | 3.46 | 7.51%
Q11 | 0.5 | 0.44 | 13.64%
Q12 | 1.78 | 1.78 | 0.00%
Q13 | 4.4 | 4.19 | 5.01%
Q14 | 0.91 | 0.97 | -6.19%
Q15 | 1.98 | 2.05 | -3.41%
Q16 | 0.97 | 0.91 | 6.59%
Q17 | 5.54 | 5.27 | 5.12%
Q18 | 7.68 | 7.42 | 3.50%
Q19 | 2.72 | 2.58 | 5.43%
Q20 | 1.51 | 1.51 | 0.00%
Q21 | 3.39 | 3.32 | 2.11%
Q22 | 0.57 | 0.57 | 0.00%

----

version info with `pgo-sample` and `fdo`

```
TiFlash
Release Version: v6.2.0-alpha-29-gec199a817-dirty
Edition:         Community
...
Enable Features: jemalloc avx avx512 unwind thinlto pgo-sample fdo
Profile:         RELWITHDEBINFO
...
```

#### Benchmark-100G

env:
- 3 tiflash, 3 pd, 3 tikv, 1 tidb
- data: tpch-100
- tiflash replica: 3
- limit tiflash cpu usage up to 10 cores(1000%) by cgroup.
- original commit: 97342db2a1dd705b1bd98f3481d51f48a4f0300d

----

Time Cost(s) | LTO | FDO+LTO | FDO+LTO : LTO
-- | -- | -- | --
Q1 | 12.08 | 11.04 | 9.42%
Q2 | 4.33 | 4.26 | 1.64%
Q3 | 8.22 | 8.09 | 1.61%
Q4 | 21.17 | 21.64 | -2.17%
Q5 | 19.36 | 19.83 | -2.37%
Q6 | 1.91 | 1.85 | 3.24%
Q7 | 9.83 | 9.97 | -1.40%
Q8 | 11.98 | 11.58 | 3.45%
Q9 | 65.26 | 64.32 | 1.46%
Q10 | 10.57 | 10.37 | 1.93%
Q11 | 2.05 | 2.05 | 0.00%
Q12 | 5.2 | 5.27 | -1.33%
Q13 | 12.72 | 12.11 | 5.04%
Q14 | 2.18 | 2.11 | 3.32%
Q15 | 4.13 | 4.06 | 1.72%
Q16 | 2.32 | 2.25 | 3.11%
Q17 | 18.29 | 18.22 | 0.38%
Q18 | 23.12 | 22.85 | 1.18%
Q19 | 5.87 | 5.74 | 2.26%
Q20 | 4.06 | 4.13 | -1.69%
Q21 | 39.02 | 39.29 | -0.69%
Q22 | 1.38 | 1.31 | 5.34%

#### Conclusion

Except Q1, io/net/schedule-wait cause much more latency in other queries when using multi nodes with large data. We believe if usage about cpu resource is relatively high, the improvement brought by fdo will be better.

### Check List

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- bugfix or new feature needs a release note -->

```release-note
None
```