@solotzg 2022-07-11
Add benchmark about tpch100 with 3 tiflash nodes. PTAL @zanmato1984 .